Migrate Soldier check to wallet store
