+++ src/store/wallet/mutations.js
@@ -1,3 +1,5 @@
+import { NFT } from "@/data/NFTs";
+
 export default {
   setOnChainBalance: (state, balances) => {
     state.onChainBalance = balances;
@@ -5,4 +7,14 @@ export default {
   setOffChainBalance: (state, balances) => {
     state.offChainBalance = balances;
   },
+  setSoldierLoading(state, loading) {
+    state.isSoldierLoading = loading;
+  },
+  setSoldier(state, { type, soldier }) {
+    if (type === NFT.HUMAN) {
+      state.humanSoldier = soldier;
+    } else {
+      state.orcSoldier = soldier;
+    }
+  }
 };
