+++ src/store/war/actions.js
@@ -1,6 +1,4 @@
 import WarsController from "@/controller/WarsController";
-import SoldierController from "@/controller/SoldierController";
-import { NFT } from "@/data/NFTs";
 
 export default {
   async getWar({ state, commit, dispatch }, warId) {
@@ -24,7 +22,7 @@ export default {
       dispatch("checkWarCountdown");
       dispatch('getAccountPrizes', currentWar.id);
       await dispatch("user/fetchUserWallet", null, { root: true });
-      await dispatch("checkSoldiers");
+      await dispatch("wallet/checkSoldiers");
       dispatch("checkCompleteFormations");
       dispatch("enlistment/checkPlayerEnlistment", null, { root: true });
     } catch (error) {
@@ -137,23 +135,5 @@ export default {
       }
     }
     commit('setCompleteFormations', completeFormations);
-  },
-  async checkSoldiers({commit, rootState}) {
-    const controller = new SoldierController();
-    commit('setSoldierLoading', true);
-    try {
-      const soldier = await controller.getNFTByType(rootState.user.account, NFT.HUMAN);
-      console.log(soldier)
-      commit('setSoldier', {type: NFT.HUMAN, soldier });
-    } catch (error) {
-      console.error(error);
-    }
-    try {
-      const soldier =  await controller.getNFTByType(rootState.user.account, NFT.ORC);
-      commit('setSoldier', {type: NFT.ORC, soldier });
-    } catch (error) {
-      console.error(error);
-    }
-    commit('setSoldierLoading', false);
   },
 };
