+++ src/store/war/index.js
@@ -22,9 +22,6 @@ const initialState = () => ({
     isLoading: false,
     data: null,
   },
-  isLoadingSoldier: false,
-  humanSoldier: null,
-  orcSoldier: null,
   countdown: 0,
   phase: 'not-started',
   accountPrizes: [],
